// src/models/service.ts
export type Servicio = {
    id_servicio: number;
    id_usuario: number;
    nombre: string;
    enlace: string;
    descripcion: string;
    estatus: number;
    orden: number;
    costo: number;
    estatus_web: number;
    estatus_sistema: number;
    fecha_registro: string;
    fecha_eliminado: string;
}